fixed compile error on POSIX
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/trunk@3665 212acab6-be3b-0410-9dea-997c60f758d6

+++ Source/script.cpp
@@ -4421,7 +4421,7 @@ int CEXEBuild::doCommand(int which_token, LineParser &line)
 
           if (ver)
           {
-            if (versize > sizeof(WORD) * 3)
+            if ((size_t) versize > sizeof(WORD) * 3)
             {
               // get VS_FIXEDFILEINFO from VS_VERSIONINFO
               WCHAR *szKey = (WCHAR *)(ver + sizeof(WORD) * 3);
